requirements.txt
Flask~=3.0.0
tensorflow-cpu==2.10
tensorflow-directml-plugin
numpy~=1.26.1
opencv-python~=4.8.1.78